+++ sklearn/neural_network/__init__.py
@@ -3,5 +3,4 @@ The :mod:`sklearn.neural_network` module includes models based on neural
 networks.
 """
 
-from .rbm import BernoulliRBM
-from .gaussian_bernoulli_rbm import GaussianBernoulliRBM
+from .rbm import BernoulliRBM, GaussianBernoulliRBM
